AppCha 패키지, 테마 변경
