+++ presentation/.gitignore
